Move background repairs scheduling to diag.py
schedule_background_repairs() now lives in ops/diag.py alongside the
repair functions it calls. main.py's _maybe_schedule_background_repairs
becomes a thin guard (startup class + bootstrap check) that delegates.

main.py: 1131 → 1092 lines. Removes 2 pyright-ignore comments for
private-usage imports.
